// frontend/src/concepts/modelRegistry/types.ts
export type ModelRegistryBase = {
  id?: string;
  name?: string;
  externalID?: string;
  description?: string;
  createTimeSinceEpoch?: string;
  lastUpdateTimeSinceEpoch?: string;
  customProperties?: Record<string, unknown>;
};
// ...
export type ModelRegistryListParams = {
  size: number;
  pageSize: number;
  nextPageToken: string;
};

export type RegisteredModelList = ModelRegistryListParams & { items: RegisteredModel[] };
export type ModelVersionList = ModelRegistryListParams & { items: ModelVersion[] };
export type ModelArtifactList = ModelRegistryListParams & { items: ModelArtifact[] };
export type ModelRegistryListParamsOptions = Omit<ModelRegistryListParams, 'nextPageToken'> & {
  page?: number;
};

export type ModelRegistryListPaged<T extends ModelRegistryBase> = ModelRegistryListParams & {
  items?: T[];
};